tests: remove obsolete framework.py
rowcache_invalidator was the last one using it for MultiDict.
Moved that class into the test itself.
